aarch64: Enable Windows on Arm builds
- make necessary changes for building oneDNN with MSVC and Ninja
- protect build with (basic) CI

We do not build with ACL as ACL can not be built with MSVC (it requires Clang), and we have not yet fixed Clang builds for oneDNN

+++ src/cpu/aarch64/jit_uni_binary_kernel.cpp
@@ -432,7 +432,7 @@ void jit_uni_binary_kernel_t<isa>::compute_bcast(bool tail) {
 
 template <cpu_isa_t isa>
 void jit_uni_binary_kernel_t<isa>::push(const Xbyak_aarch64::XReg &reg) {
-    str(reg, pre_ptr(X_SP, -(reg.getBit() / 8)));
+    str(reg, pre_ptr(X_SP, -static_cast<int>(reg.getBit() / 8)));
 }
 template <cpu_isa_t isa>
 void jit_uni_binary_kernel_t<isa>::pop(const Xbyak_aarch64::XReg &reg) {
